Add enum ActionEvent.Type so all actionevents no longer look alike.
Added an enum with multiple subtypes of action event, and use them
everywhere by adding a "type" to all the constructors that previously
had many different purposes.   For example, this makes it possible to
distinguish a "pointerPressed" event from a "pointerReleased" event.
The original, untyped constructors are still present to preserve
compatibility with client code that already uses ActionEvents.