Performance improvements to Androids graphics
Minimised the overhead of drawString in a similar way to the iOS port
by drawing on a bitmap within the EDT instead of on the Android thread.
Drawing text is just really damn slow on Android. Eliminated as much
overhead for calls as possible and added some commented out debugging
logic to make future iterations easier. Some jank still appears but not
nearly as much as before.