Fixed BackgroundFetch support in android so that, at least, it will not crash.  Unfortunately, this API cannot be implemented in a way that allows the background fetch to be performed in the background always.   If the application has been unloaded and the display deinitalized, then this will start the app into the foreground again for the background fetch to resume.  If the app has just been paused, then it will correctly run the background fetch in the background.  This behaviour is consistent with background location updates on Android.  It also means that we don't have to worry about murky edge cases where the app is opened while a fetch is running in the background.
While this implementation seems to be working and well-understood, the behaviour of opening the app into the foreground is not good -- the app would be viewed as a bad citizen.  Unfortunately there doesn't appear to be a way to get around this.  The CodenameOne Android implementation is too closely tied to an Activity.  If one wants to provide background fetch support, it would have to be via either a separate limited API that is accissible on the background, or via a more static API that doesn't actually allow code to run on the background, but rather sets up some tasks to run with parameters, and they are run in native space periodically.

Either way, this will need to be addressed later.

This commit, delivers at least a working BackgroundFetch implementation as promised for the 3.5 release.