Decoupled Android port from Activity so that Codename One can be initialized from a Service without the need for an Activity.  This fixes issues with background APIs like background fetch and background location so that they no longer cause the app to come to the foreground when a background process is started.
This change may look like it is far reaching but it is very simple.  I encapsulated the AndroidImplementation.activity property into a getter that will return the activity if it is set, or null otherwise.  I also added a getContext() method that will always return a context.  This will be either a Service or the standard CodenameOneActivity.  I then replaced all references to 'activity' with getContext() if an Activity wasn't strictly required.  If an activity was strictly required, I would check if getActivity() returns null first, and fail if it does.

Now, if you try to init cn1 inside a service, you can pass the service to Display.init(service), and that service will be treated like the context.  If the user opens the app while a service is running, it will call Display.init() with the CodenameOneActivity which will replace the context with itself seamlessly.

The CodenameOneBackgroundLocationActivity and CodenameOneBackgroundFetchActivity classes are now deprecated.  They should be removed --- but I didn't want to do that just yet.