Changed android port so that it will try to use resizing rather than panning every time we activate the native editor, and switch to panning if the field is covered by the keyboard.  This is to fix this bug.  https://github.com/codenameone/CodenameOne/issues/1827
I made an attempt to fix this directly by wrapping the textview in a scrollview so that the text field would be fully scrollable even if its cn1 peer was pushed off the top of the screen.  However, I couldn't get it to cover all edge cases.  For example, if the user tries to select some of the text, Android will implicily force it to the resize edit mode without providing any cues which causes problems with the padding in the scroll view.  With some additional effort, it might be possible to proceed with this approach, but the solution I arrived upon should sufficiently reduce the incidences of the text area scrolling bug that it is a non-issue - and it appears to be a robust solution.

The gist of the previous attempt can be seen at https://gist.github.com/shannah/471033878e53c1fc297680fa85f6fd20