Implemented peer-z-layering in iOS.  https://github.com/codenameone/CodenameOne/issues/1758 This also puts in place a structure for other platforms to more easily support peer z-layering, as it adds the concept of a front graphics layer that is drawn on top of peer components.  Graphics now has a flag 'enableFrontGraphics' which, when turned on, causes drawing operations to be buffered as soon as a peer component is encountered so that they can be passed to the front graphics buffer when the graphics is flushed.
Any platform that overrides their implementation's isFrontGraphicsSupported() to return true, will set this graphics flag on its main graphics context.  There are a few other methods that also need to be implemented in the implementation to support front graphics, see the ios implementation in this commit for an example.

Front graphics can be enabled or disabled at runtime by setting the useFrontGraphics display property to 'true' or 'false'.