Added ability to perform action on ToastBar tap
There are cases when ToastBar is used as an in-app notification and we
want to open up another form or perform other actions.

The extra action is added to an existing leadButton and doesn't affect
hiding of the ToastBar when it's touched.

Here is a simple example of usage:
`ToastBar.showMessage("You have a message from Diamond, click here to
view.", FontImage.MATERIAL_CHECK_CIRCLE, 5000, e -> performAnAction());`