eliminate the explicit list of threads due for garbage collection
Eliminate the threadsToDelete array and associated bookkeeping.
Instead, finish the thread off when it actually dies, and depend on
the finalizer to eventually call back for final disposal.  Note that
this uses the address of the private thread structure as the public
thread id instead of the unix thread number.  If that's deemed
undesirable for some reason, we could instead add a new private member
to the Thread class.