fix the "expected statement" bug in ios builds
Use the label itself instead of a string representing the label
as the key for duplicate label detection.  This prevents the "expected
statement" bug that results when an unreferenced label is left at the
end of a method.

Also add a cleanup phase so duplicate runs (during debugging for
example) will start with empty lists of used labels and parsed classes.