Added signal handler to convert EXEC_BAD_ACCESS into NPEs that can be caught.  Other signals that previously would crash an app are converted to RuntimeExceptions and thrown - and can be caught. Note: When using the Xcode debugger, these exceptions won't be thrown - it will simply break at the signal location, as it would previously.  However, when running on device, this should eliminate a whole class of mysterious crashes. Due to this channge I recommend leaving the ios.fieldNullChecks build hint set to false, as NPEs will be caught anyways, when a signal is received.
So far, testing seems to show that the app behaves normally after catching an exception that originated from a BAD_ACCESS signal, however it is possible, even likely, that there are some negative side-effects - so it isn't a good idea to rely on this too much in an app.  E.g. don't do a lot of try/catch NPE routinely in an app.  Fix NPEs when they are found.
https://github.com/codenameone/CodenameOne/issues/2135