Added support for 'inline' styles which take precedence over UIID styles.  Inline styles are to UIIDs in CN1 as the 'style' attribute is to CSS classes in HTML.  Inline styles are applied to components as strings which follow a specific format that is designed to work well with the GUI builder.  The style strings are parsed at runtime to generate Styles dynamically.  Inline styles are cached as normal UIIDs with a name derived using both the component UIID and the component name.  Caching the styles in the theme should make performance hits negligible.
Syntax for inline style strings is similar to CSS, but it is important to note that it is *NOT* CSS.  Property names and values correspond closely with the properties of the com.codename1.ui.plaf.Style class.  More documentation is forthcoming on this syntax.  A few example style strings:
fgColor:ff0000; bgColor:0000ff; border:1px solid ff0000; font:18mm native:MainRegular; transparency:255;
border:splicedImage /bg.png 0.4 0.4 0.4 0.4

The splicedImage example points to another change in this comment: the addition of a the Border.createImageSplicedBorder() method, which creates a 9-piece image border at runtime using only a single image, and splice insets, which are used to splice the image into 9 sub-images.

This commit is part of a larger change to the GUI builder that simpifies the process of styling components.