Fixed formLayeredPane not playing nice with pointer events.  This is an alternate fix for the same problem targeted by previous commits: https://github.com/codenameone/CodenameOne/commit/2b94acaf37d3e7c9569be57b1bd577142af08051 https://github.com/codenameone/CodenameOne/commit/e0ca7321a650b0018a5ece901554288d42efb6f9 https://github.com/codenameone/CodenameOne/commit/fe0aedb9209c3daaf37d391e26ab12e62c0cf5ce
But the previous attempt introduced some regressions, and the changes in Container.getComponentAt(x,y) and Form.getActualPane() were reverted:
https://github.com/codenameone/CodenameOne/commit/e152cbe6defec038772a91214ef039c90a1572a0

This new fix does not attempt to change Container.getComponentAt(x,y) or Form.getActualPane() anymore as these methods have a lot of legacy baggage and need to continue to work the way they do, even if that functionality isn't well-defined or consistent.  Instead, I have explicitly checked the formLayeredPane in every instance that getActualPane() is called, to determine if the formLayeredPane should intercept the functionality.  This circumvention was required for pointer events, component focusing, and for dragging.

Also changed some usage in the JavaSE, iOS, and Android ports to use the new, well-defined getResponderAt(x,y) method instead of getComponentAt(x,y) since the uses in those locations assumed functionality in line with getting a responder - and not just any component.

Two key public methods were added to the API to support this behaviour:

1. Container.getResponderAt(x, y) - gets the topmost component that responds to pointer events.  This includes focusable, and scrollable components, as well as components that are marked to grab pointer events.
2. Component.respondsToPointerEvents() - Checks to see if a component responds to pointer events.  This includes focusable, scrollable, and isGrabPointerEvents() components that are enabled and visible.

Peer components are now set to grab pointer events by default.  If you have a peer component that you don't want to grab pointer events, you can call setGrabPointerEvents(false) on it.