Static method to convert a Map to a valid JSON string.
I noted that the CodenameOne API is missing of a method to convert a Map<String,Object> to a JSON string, that's why I created a static method to convert the given Map<String,Object> to a valid JSON representation. The values allowed types in the Map<String,Object> are: Number, String, Boolean, List, Map or null. Limited whitespace is inserted be make the resulting JSON string more readable. Simple and complex examples of usage are in the javadoc. I tested the generated JSON strings in a JSON validator. I hope that my proposed method will be helpful.