Fixed #2189 but went quite a bit beyond that
The gist of this commit allows you to specify HTTP header values for the browser component on Android & iOS. So you could just set the accept-language header to whatever you want in the setURL() call. Naturally this should work (in theory) for all types. Unfortunately this isn't doable in the pathetic JavaFX browser component so we can't simulate it.