Main focus of this commit is to add a lightweight mode flag to the Picker class which causes it to prefer lightweight widgets rather than native widgets.  It includes a new set of Spinner3D widgets (which aren't yet publicly exposed) for the lightweight pickers.  These Spinner3D classes are built on top of a new scene graph implementation (in the com.codename1.ui.scene package) which is currently marked deprecated for internal use only.  The scene graph allows us to create advanced 3D widgets.
While the Spinner3D classes will ultimately support full perspective (like the native iOS picker), they are currently juse emulating perspective using 2D transforms.  The code is mostly complete for full perspective support, but debugging on iOS has been difficult, so I have left it turned off for now.

In lightweight mode, all of the picker types are supported except the duration pickers.  Those will be implemented soon as well.

There are still things that need to be done to make this a full viable replacement for native pickers.  For example, we need to add the ability to make the various competing input forms cooperate.  E.g. If the picker is opened, and the user starts editing a text field, then the picker should close automatically as the keboard opens.  And vice-versa.

Another rough edge that needs to be smoothed out is the case where the picker covers dialog covers up the field that opened the picker.  It should automaticlaly scroll the forum so that the field is visible, but it doesn't do this yet.

Tablet support is still todo also.

Most of these todo items will be committed in the next few days.