Added initial support for rich push notifications.  This commit includes support in android and iOS for attaching images to push notifications, but it also includes all of the groundwork necessary to expand this to other features of rich push notifications, like actions. This also includes some improvements to the push simulator to support all of the different push notification types. This commit is only part of a series of commits to various projects for adding support for rich push notifications.  The build server, certificate wizard, Javascript port, and push server all require updates - commits forthcoming for those projects.
Support for push notifications required a new push type: 99
The message body for push type 99 is an XML snippet of the form:
<push type='X' body='thebody' category='category'><img src='https://example.com/image.jpg'/></push>

The type and body attributes correspond with the push type of the main push content.  In this way, all of the existing push types can be expressed using this structure, and they can all have rich content added to them.  E.g.

<push type='1' body='Hello World'/>

Would be the same as just sending a message with body 'Hello World' and type=1.