Fixed #2493 which is really a bug
There was no theme applied to the desktop skin causing it to pick remainders from the last skin that was running. Made it use iOS theme which is more consistent. We should probably have more options but since the desktop & web versions use iOS by default this makes more sense.