Fixed issue with select menus not showing up in webviews on iPad.  https://github.com/codenameone/CodenameOne/issues/2571
This turned out to be an issue with the way we were doing Z-Layering.  When adding a peer component, we were adding a peer components UIView to the parent of the root view.  Unfortunately, this technically places it outside the view hierarchy - even though it appeared to work.  Resolved the problem by instead re-rooting the view hierarchy with a new UIView above the previous EAGLView.  There were some places where the view controller (CodenameOne_GLViewController) would call self.view and expect that it would be the EAGLView.  This can no-longer be assumed.  If there have been any peer components added, then self.view will return the new UIView root (the parent of the EAGLView).

To help with this, I added a new eaglView method in CodenameOne_GLView controller which will always return the EAGLView.  I have swapped out all of the places where an EAGLView was explicitly assumed via cast.  It is possible that there are other situations where it is implicitly assumed, but we'll have to squash those as we find them.

Also added a macro EAGLVIEW for [[CodenameOne_GLView Controller] instance] eaglView] for convenience.

Tested with a few different apps in test suite so far and appears to not have broken anything - but this is a pretty low-level change so it could have far-reaching effects.