Added setRenderingHint() method to Graphics class to pass rendering hints to the native graphics.  Currently only one value is supported: Graphics.RENDERING_HINT_FAST.  And only iOS uses this when rendering images.  Since copying images to textures is expensive, we keep the last generated texture cached.  This works well if we are always rendering the image at the same size.  If we are constantly rendering the same image at different sizes, then we'll constantly be invalidating the cache - which affects UX. This affected pinch zoom in the image viewer causing it to be choppy as it had to regenerate a texture for every change in size. When Fast rendering is enabled, we now only invalidate the texture cache if the image is larger or smaller than the existing texture by more than a factor of 2.
Also fixed a bug that caused images to be rendered as black if they are larger than the max OGL texture size.  Now it will cap the size of the texture at the max OGL size, and it will use the GPU to scale the texture to the desired size.
This fixes https://github.com/codenameone/CodenameOne/issues/2786