Commented out location manager map
Unfotunately google cracked down on using the API key without billing information. This puts us at risk of theft for this feature so it would need to be rewritten and I'm not sure if that's technically feasible. A better approach would probably be dumping google maps for OSM or similar. But I'm not sure if that's feasible. For now it's commented out.