Created interfaces IconHolder and SelectableIconHolder to be implemented on Label-like and Button-like components that can have icons.  This allows the manipulation of components' icons in a more generic way.  IconHolder includes methods like setIcon(), getIcon(), etc... .  SelectableIconHolder includes methods for managing icons of the various states, e.g. setDisabledIcon(), etc... Label and SpanLabel implement IconHolder.  Button, MultiButton, SpanButton, SpanMultiButton implement SelectableIconHolder.
Added support for styling icons differently than text in IconHolder components (e.g. Labels, Buttons, etc..).  This allows us to use a gray icon with black text, or to make the icon font size larger than the text.  The IconHolder interface setIconUIID(String) method is used for this purpose.  The icon styles can be obtained using the IconHolder getIconStyleComponent() which returns a component whose styles can be used for styling the icon.  By default, this method will return the IconHolder itself (e.g. the Button or Label itself), but in cases where setIconUIID(String) has been called, it will return a separate dummy component whose styles can be used.  FontImage has been adjusted to use these methods when creating icons for labels.

Additional methods have been added to FontImage for setting font icons on IconHolders.

Added support for state change events on Buttons (Button.add/removeStateChangeListener()).  Used this to implement Button#bindStateTo(Button) and Button#unbindStateFrom(Button) which allows a button to bind its state to another button, so when that other button is pressed, this button will enter the pressed state, etc..  This was necessary to implement state-ful icon methods in MultiButton since the icon is not the lead component, but we want to be able to add disabled buttons, and pressed buttons for them.  In MultiButton, the 'icon' component binds to the 'emblem' (which is the lead component) so that when the emblem is pressed, the 'icon' will display its 'pressed' icon, etc...