Efficiency improvements for Container.revalidate().  Added method Container.revalidateLater() which works like revalidate() except that the revalidation is deferred until just before the next EDT paint cycle.  This prevents multiple reflows of the UI between paints.  revalidateLater() will prevent duplicate revalidations by ensuring that each container scheduled for a revalidation is only revalidated once.  If a container's parent container is scheduled for revalidation, then it will not be revalidated. Changed several existing method calls to revalidateWithAnimationSafety() to revalidateLater().
Added display property 'Form.revalidateFromRoot' which, when set to 'false', will disable the current behaviour that calls to revalidate() will trigger a revalidate on the parent form also.  Not sure why this is the behaviour, and I think it would make sense to eventually default this flag to 'false' (it is currently default 'true') so that revalidates aren't so aggressive.