Removed JavaFX dependency from the CSS compiler.  The CSS compiler now uses the BrowserComponent for producing snapshots which should work with both JavaFX and CEF.  It has the added benefit that, at least with CEF, it no longer needs to open a popup window to generate 9-piece borders in the browser. NOTE: This update depends on a forthcoming update to CEF which adds transparency support.  IF you run this directly right now, the 9-piece borders generated with CEF won't include any transparency which will just plain look bad in cases where it needs transparency. NOTE2: Haven't tested yet with JavaFX.
NOTE3: For speed and efficiency, the CSS compiler will try to compile the CSS without loading CEF onto the classpath.  If it runs into CSS that requires taking a snapshot in a browser component, and CEF and JavaFX are not loaded, then it will load CEF onto the classpath (or attempt to), and try to compile again.