fix: make gradle dependencies use implementation instead of compile
compile is deprecated in gradle 7+, and users have had to explicitly add the android.arrimplementation build hint for the ZBarScanner library to force it to include via implementation.
This change makes implementation the default behaviour when using gradle 6+, but with the ability to override for a particular library using the new android.arrcompile build hint
which works the same way as the android.arrimplementation build hint.