Update README.md
Here I have added a link to w3schools.com The page that will be brought up is a guide on using CSS. In the section "CSS Live Update" it shows how CSS can be used to enhance the app's look. However it does not provide any further information on how to use the different features of CSS or where to look. 
This change can allow users to find a place to look up ways they can enhance their website along with guides, and visuals of how to code it.