fix: [3914] Fix safe area insets on android 34
This fixes both safe area insets and form bottom padding editing on android 34.
By trial and error, I found that, with respect to immersive mode, android 34 behaves like
android 33 and earlier unless we explicitly enable it by calling WindowCompat.setDecorFitsSystemWindows(window, false),
which we don't do.  And it is not possible to detect whether this has been called.
This patch adjusts immersive heuristics to check for API 35+ instead of 34+ when determining if we are drawing
edge to edge on the screen.

This impacts the safe area insets implementation as well as form bottom padding editing.

As part of this, I have re-enabled the android.useSafeAreaInsets display property by default
so taht we use safe area insets by default.  This can be opted out.

The reason is that on API 35+ it needs this setting to work properly.